Complete fifth flex exercise
I thought I needed a flex container around the buttons, but those are
inline elements so I didn't. It made sense to me to split the content
into a left-container and a right-container, though I see that I could
have left the icon on its own. Though flex-shrink: 0 wasn't needed
because it was wrapped in a div. The top-line div was added to split the
text and the x. A bit more intuitive than the Odin solution.